refactor: HelpPanel이 불필요하게 렌더링되지 않도록 구조 개선
