+++ src/pages/Search/Search.tsx
@@ -1,31 +1,12 @@
-import { KeyboardEvent } from 'react';
-import useGifSearch from './hooks/useGifSearch';
-
-import SearchBar from './components/SearchBar/SearchBar';
-import SearchResult from './components/SearchResult/SearchResult';
+import SearchSection from './components/SearchSection/SearchSection';
 import HelpPanel from './components/HelpPanel/HelpPanel';
 
 import styles from './Search.module.css';
 
 const Search = () => {
-  const { status, searchKeyword, gifList, searchByKeyword, updateSearchKeyword, loadMore } =
-    useGifSearch();
-
-  const handleEnter = (e: KeyboardEvent<HTMLInputElement>) => {
-    if (e.key === 'Enter') {
-      searchByKeyword();
-    }
-  };
-
   return (
     <div className={styles.searchContainer}>
-      <SearchBar
-        searchKeyword={searchKeyword}
-        onEnter={handleEnter}
-        onChange={updateSearchKeyword}
-        onSearch={searchByKeyword}
-      />
-      <SearchResult status={status} gifList={gifList} loadMore={loadMore} />
+      <SearchSection />
       <HelpPanel />
     </div>
   );
